+++ resources/ReleaseNotes.md
@@ -9,3 +9,4 @@
 - Make sure converted values in `mag.process.SignedInteger` are returned as `double`
 - Fix [#10](https://github.com/ImperialCollegeLondon/MAG-Data-Visualization-Toolbox/issues/10)
 - Fix [#19](https://github.com/ImperialCollegeLondon/MAG-Data-Visualization-Toolbox/issues/19)
+- Add tests for processing step documentation properties